// src/ts/tmwindow/TmWindow.d.ts
import { ITmWindowOptions } from "./ITmWindowOptions";
export default class TmWindow {
    readonly domElement: HTMLElement;
    private readonly options;
    private headerElement;
    private titleElement;
    private contentElement;
    private lastStyle;
    private mouseDownEv;
    constructor(options?: ITmWindowOptions | string);
    readonly isOpen: boolean;
    readonly isClosed: boolean;
    readonly isMinimized: boolean;
    title: string;
    width: number;
    x: number;
    y: number;
    content: any;
    setOption(name: string, value: any): this;
    getOption(name: string): any;
    appendElement(element: HTMLElement): void;
    setPosition(x: number, y: number): void;
    open(): void;
    close(event: any): void;
    destroy(): void;
    minimize(): void;
    reappend(): void;
    private _buildWindow;
    private _buildHeader;
    private _buildHeaderButtons;
    private _buildButton;
    private _buildContent;
    private _addRepositionEvent;
    private _repositionEvent;
}